requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pymongo==4.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
requests==2.32.2
sentence-transformers==2.6.1
langdetect==1.0.9
bcrypt<4.1
pydantic[email]
python-multipart
huggingface_hub[hf_xet]
python-dotenv
pdfplumber